+++ packages/oauth2/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.1.2 - 2020-09-03
+
+- Added `node-fetch`.
+- Fixed `getTokensFromAuthorizationCode` returned data and type.
+
 ## 0.1.1 - 2020-09-03
 
 - Added `state` to authorization_url.